requirements.txt
audioop-lts
huggingface_hub==0.25.2
torch==2.4.1
numpy<2
Pillow
opencv-python-headless
onnxruntime
transformers
accelerate
ultralytics
gradio==4.44.1